piano/requirements.txt
python-rtmidi==1.2.1
midiutil==1.2.1
slackclient==1.3.0
midi2audio==0.1.1
mido==1.2.9
google-api-python-client==1.7.5
google-auth-httplib2==0.0.3
google-auth==1.6.1
